+++ docs/reference/release-notes.md
@@ -3,12 +3,12 @@ Sigrid release notes
 
 SIG uses [continuous delivery](https://en.wikipedia.org/wiki/Continuous_delivery), meaning that every change to Sigrid or the underlying analysis is released once our development pipeline has completed. On average, we release somewhere between 10 and 20 times per day. This page therefore doesn't list every single change, since that would quickly lead to an excessively long list of small changes. Instead, this page lists Sigrid and analysis changes that we consider noteworthy for the typical Sigrid user.
 
-### January 6, 2025
+### January 5, 2025
 
 - Happy new year from everyone at SIG!
 - **Sigrid CI:** Sigrid CI for Open Source Health is no longer in beta, and is now available to all Sigrid users with an Open Source Health license. This documentation includes [instructions for enabling Open Source Health in your existing Sigrid CI configuration](../sigridci-integration/using-sigridci.md#adding-open-source-health-feedback-to-an-existing-sigrid-ci-configuration).
-- **Sigrid CI:** Open Source Health feedback has been extended to now include feedback on both vulnerability risk and license risk. The latter will only apply if you have defined a [license risk objective](../capabilities/portfolio-objectives.md).
 - **Sigrid CI:** Open Source Health feedback now explicitly marks transitive dependencies, making it easier to differentiate them from direct dependencies.
+- **Sigrid CI:** Open Source Health feedback has been extended to now include feedback on both vulnerability risk and license risk. The latter will only apply if you have defined a [license risk objective](../capabilities/portfolio-objectives.md).
 
 ### December 15, 2025